Move ConsoleContextMenu to a separate file
