+++ Sources/PulseUI/Features/Console/ConsoleView-ios.swift
@@ -9,14 +9,9 @@ import Combine
 
 #if os(iOS)
 
-import UniformTypeIdentifiers
-
 public struct ConsoleView: View {
     @ObservedObject var viewModel: ConsoleViewModel
     @State private var isSharing = false
-    @State private var isShowingSettings = false
-    @State private var isShowingStoreInfo = false
-    @State private var isDocumentBrowserPresented = false
 
     public init(store: LoggerStore = .shared) {
         self.viewModel = ConsoleViewModel(store: store)
@@ -39,7 +34,7 @@ public struct ConsoleView: View {
                 trailing: HStack {
                     ShareButton { isSharing = true }
                     if #available(iOS 14.0, *) {
-                        contextMenu
+                        ConsoleContextMenu(store: viewModel.store)
                     }
                 }
             )
@@ -52,65 +47,6 @@ public struct ConsoleView: View {
                     ShareView(ShareItems(messages: viewModel.store))
                 }
             }
-            .sheet(isPresented: $isShowingSettings) {
-                NavigationView {
-                    SettingsView(store: viewModel.store)
-                        .navigationBarItems(trailing: Button(action: { isShowingSettings = false }) {
-                            Text("Done")
-                        })
-                }
-            }
-            .sheet(isPresented: $isShowingStoreInfo) {
-                if #available(iOS 14.0, *) {
-                    NavigationView {
-                        StoreDetailsView(source: .store(viewModel.store))
-                            .navigationBarItems(trailing: Button(action: { isShowingStoreInfo = false }) {
-                                Text("Done")
-                            })
-                    }
-                }
-            }
-            .backport.fullScreenCover(isPresented: $isDocumentBrowserPresented) {
-                if #available(iOS 14.0, *) {
-                    DocumentBrowser()
-                }
-            }
-    }
-
-    @available(iOS 14.0, *)
-    private var contextMenu: some View {
-        Menu {
-            Section {
-                Button(action: { isShowingStoreInfo = true }) {
-                    Label("Store Info", systemImage: "info.circle")
-                }
-                if !viewModel.store.isArchive {
-                    Button(action: { isDocumentBrowserPresented = true }) {
-                        Label("Browse Stores", systemImage: "doc")
-                    }
-                }
-            }
-            Section {
-                Button(action: { isShowingSettings = true }) {
-                    Label("Settings", systemImage: "gear")
-                }
-            }
-            if !viewModel.store.isArchive {
-                Section {
-                    if #available(iOS 15.0, *) {
-                        Button(role: .destructive, action: buttonRemoveAllTapped) {
-                            Label("Remove Message", systemImage: "trash")
-                        }
-                    } else {
-                        Button(action: buttonRemoveAllTapped) {
-                            Label("Remove Message", systemImage: "trash")
-                        }
-                    }
-                }
-            }
-        } label: {
-            Image(systemName: "ellipsis.circle")
-        }
     }
 
     private var contentView: some View {
@@ -127,22 +63,6 @@ public struct ConsoleView: View {
         if viewModel.entities.isEmpty {
             PlaceholderView.make(viewModel: viewModel)
         }
-    }
-
-    // MARK: Helpers
-
-    private func buttonRemoveAllTapped() {
-        viewModel.store.removeAll()
-
-#if os(iOS)
-        runHapticFeedback(.success)
-        ToastView {
-            HStack {
-                Image(systemName: "trash")
-                Text("All messages removed")
-            }
-        }.show()
-#endif
     }
 }
 
@@ -173,17 +93,6 @@ private struct ConsoleToolbarView: View {
                 ConsoleFiltersView(viewModel: viewModel.searchCriteria, isPresented: $isShowingFilters)
             }
         }
-    }
-}
-
-@available(iOS 14.0, *)
-private struct DocumentBrowser: UIViewControllerRepresentable {
-    func makeUIViewController(context: Context) -> DocumentBrowserViewController {
-        DocumentBrowserViewController(forOpeningContentTypes: [UTType(filenameExtension: "pulse")].compactMap { $0 })
-    }
-
-    func updateUIViewController(_ uiViewController: DocumentBrowserViewController, context: Context) {
-
     }
 }
